requirements.txt
crewai>=0.5.0
langchain>=0.1.0
openai>=1.3.0
faiss-cpu>=1.7.4
fastapi>=0.104.0
uvicorn>=0.23.2
pydantic>=2.4.0
python-dotenv>=1.0.0
PyYAML>=6.0.1
pytest>=7.0.0
black>=23.3.0
isort>=5.12.0
flake8>=6.0.0
pytest-mock>=3.10.0
pytest-cov>=4.1.0
httpx>=0.24.0  # For TestClient in FastAPI tests 